Refactor: Move misc havok classes
