+++ src/StudioCore/Havok/NewHavokAnimation.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Numerics;
 
-namespace StudioCore;
+namespace StudioCore.Havok;
 
 public abstract class NewHavokAnimation
 {
